Create Dnd_manager
This is what I've been tinkering around during my studies to learn Python. Started as a method to practice certain functions and iterations and I guess it still is that. But also motivated to keep improving it. 

It's the work of a beginner looking to learn. Any feedback will be greatly appreciated.